working version
contains three scripts:
- mutations.py
- reads.py
- vcf_compare.py

mutations is now working as desired - will read in a .fasta file and apply mutations to a genome (0-20 indels and 300 SNPs). The file is outfile as a text file in .fasta layout called 'mutated_sequence.txt' and a change log called 'change_log.txt'.

reads.py takes in a .fasta file and will generate a set number of reads of a certain length. Currently, this is manually set but future updates could turn this into a function where the inputs as a .fasta file and the desired avg. depth of reads and read length. reads.py outputs a .fastq file of reads.

vcf_compare takes in two vcfs and strips unnecessary information to generate two lists, which is then directly compares and returns two measures: precision (% of direct matches) and recall (% number of matches)